Add scope support to `ServiceRegistryBuilder`
